refactor(web): extract SettingsGeneral component from Settings.tsx
Moves General/Library tab content into SettingsGeneral. Settings.tsx passes
state as props. No logic changes — purely structural refactoring.
First of three PRs splitting the 4,166-line Settings.tsx.
